+++ src/middleware/csrf.js
@@ -2,18 +2,19 @@
  * Sets a request header with the value of a cookie from document.cookie, if it exists
  */
 export default (cookieName = 'csrfToken', headerName = 'x-csrf-token') => function CsrfMiddleware () {
+  const REGEXP_COOKIE_NAME = new RegExp(cookieName + '[^;]+')
+  const getCookie = () => {
+    const cookieString = REGEXP_COOKIE_NAME.exec((document || {}).cookie || '')
+    return cookieString ? decodeURIComponent(cookieString.toString().replace(/^[^=]+./, '')) : undefined
+  }
+
   return {
     request (request) {
       if (typeof document === 'undefined') {
         return request
       }
 
-      const getCookie = (cookieName) => {
-        const cookieString = new RegExp(cookieName + '[^;]+').exec((document || {}).cookie || '')
-        return cookieString ? decodeURIComponent(cookieString.toString().replace(/^[^=]+./, '')) : undefined
-      }
-
-      const csrf = getCookie(cookieName)
+      const csrf = getCookie()
 
       return !csrf
         ? request